Add images directory within recipes directory
This will serve as a place to store my recipe images so that I can
access them throughout the website. I also added
koreangroundbeefbowl.jpg which is a picture of the dish from
koreangroundbeefbowl.html. Future recipe html documents will have an
associated picture contained in the images directory and sharing the
same name.